# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Orkestra API
  version: 1.0.0
  description: |
    Orkestra is a multi-tenant internal workflow orchestration platform.

    This API currently covers:
    - System health and diagnostics
    - Workflow definition upload and browsing
    - Authenticated identity inspection
    - Platform-admin tenant management

servers:
  - url: /api

tags:
  - name: System
    description: Platform-level endpoints for health and diagnostics.
  - name: Auth
    description: Authenticated identity endpoints.
  - name: Workflow
    description: Workflow definition upload, versioning, and retrieval.
  - name: Tenant
    description: Tenant-scoped endpoints for the currently authenticated user.
  - name: Admin
    description: Platform-admin endpoints that do not require X-Tenant-Id.

paths:
  /health:
    get:
      summary: Health check
      description: Returns basic service health and optional dependency hints.
      operationId: getHealth
      tags: [System]
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheck'

  /v1/me:
    get:
      summary: Get current authenticated user
      description: Returns basic identity information derived from the authenticated JWT.
      operationId: getCurrentUser
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /v1/tenants:
    get:
      summary: List tenants for current user
      description: Lists the tenants the currently authenticated user can access.
      operationId: listMyTenants
      tags: [Tenant]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Accessible tenants returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTenantsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /v1/workflows:
    get:
      summary: List workflows
      description: Lists workflow names for the current tenant with metadata.
      operationId: listWorkflows
      tags: [Workflow]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Workflows returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkflowsResponse'
        '400':
          description: Invalid pagination parameters or missing tenant header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      summary: Upload a workflow definition file
      description: |
        Uploads a YAML workflow definition file for the current tenant.
        The service validates the DAG and creates a new immutable version
        (auto-incremented per workflow name).
      operationId: registerWorkflow
      tags: [Workflow]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterWorkflowRequest'
      responses:
        '201':
          description: Workflow version created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterWorkflowResponse'
        '400':
          description: Invalid YAML / invalid DAG / missing fields / name mismatch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Workflow name is reserved/locked (optional).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: Workflow file too large.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '415':
          description: Unsupported media type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /v1/workflows/{name}/versions:
    get:
      summary: List workflow versions
      description: Lists available versions for a workflow name (metadata only).
      operationId: listWorkflowVersions
      tags: [Workflow]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/WorkflowName'
      responses:
        '200':
          description: Workflow versions returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkflowVersionsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Workflow not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /v1/workflows/{name}/versions/{version}:
    get:
      summary: Get workflow definition by version
      description: Returns the normalized workflow definition and graph representation.
      operationId: getWorkflowVersion
      tags: [Workflow]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/WorkflowName'
        - $ref: '#/components/parameters/WorkflowVersion'
      responses:
        '200':
          description: Workflow version returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowVersion'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Workflow version not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /admin/v1/tenants:
    get:
      summary: List tenants (platform admin)
      description: Lists all tenants in the platform. Requires platform-admin access.
      operationId: listTenantsAdmin
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tenants returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTenantsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Caller is authenticated but not a platform admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    post:
      summary: Create tenant (platform admin)
      description: |
        Creates a new tenant. The caller must have platform-admin access.
        The creator may optionally be added as the first tenant admin.
      operationId: createTenant
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Caller is authenticated but not a platform admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Tenant slug already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /admin/v1/tenants/{tenantId}/memberships:
    post:
      summary: Add tenant membership (platform admin)
      description: Adds a user as a member of a tenant with a given role.
      operationId: createTenantMembership
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantMembershipRequest'
      responses:
        '201':
          description: Membership created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantMembershipResponse'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Caller is authenticated but not a platform admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Tenant or user not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Membership already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token issued by Supabase Auth.

  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor for pagination. Omit to start from the beginning.
      schema:
        type: string

    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50

    WorkflowName:
      name: name
      in: path
      required: true
      description: Workflow name.
      schema:
        type: string
        minLength: 1
        maxLength: 128

    WorkflowVersion:
      name: version
      in: path
      required: true
      description: Workflow version number.
      schema:
        type: integer
        format: int32
        minimum: 1

    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: true
      description: Active tenant identifier for tenant-scoped requests.
      schema:
        type: string
        minLength: 1
        maxLength: 128

    TenantId:
      name: tenantId
      in: path
      required: true
      description: Tenant identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128

  responses:
    UnauthorizedError:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    ForbiddenError:
      description: Authenticated caller is not allowed to access the requested tenant.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  schemas:
    HealthCheck:
      type: object
      description: Basic service health response.
      required: [status]
      properties:
        status:
          type: string
          description: Overall status.
          example: UP
        tables:
          type: array
          description: Optional list of known DynamoDB tables (dev visibility).
          items:
            type: string

    CurrentUserResponse:
      type: object
      description: Authenticated user information derived from JWT claims.
      required: [subject]
      properties:
        subject:
          type: string
          description: Stable external user identifier (for example, JWT sub).
          example: 8f2cb0d5-7d6f-4d75-b2f9-0ef0f8f2c123
        email:
          type: string
          format: email
          nullable: true
        issuer:
          type: string
          nullable: true

    RegisterWorkflowRequest:
      type: object
      required: [name, definition]
      properties:
        name:
          type: string
          description: Workflow name to register (canonical name).
          example: invoice-processing
        definition:
          type: string
          format: binary
          description: YAML file containing the workflow definition.

    RegisterWorkflowResponse:
      type: object
      description: Result of creating a new workflow version.
      required: [name, version, createdAt]
      properties:
        name:
          type: string
        version:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
          example: "2026-02-20T22:00:00Z"

    ListWorkflowsResponse:
      type: object
      description: Paginated list of workflow names for the tenant.
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowListItem'
        nextCursor:
          type: string
          nullable: true
          description: Cursor for the next page, if any.

    WorkflowListItem:
      type: object
      description: Workflow metadata used for listing.
      required: [name, latestVersion, updatedAt]
      properties:
        name:
          type: string
        latestVersion:
          type: integer
          format: int32
          minimum: 1
        updatedAt:
          type: string
          format: date-time

    ListWorkflowVersionsResponse:
      type: object
      description: Version history for a workflow (metadata only).
      required: [name, versions]
      properties:
        name:
          type: string
        versions:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowVersionMetadata'

    WorkflowVersionMetadata:
      type: object
      required: [version, createdAt]
      properties:
        version:
          type: integer
          format: int32
          minimum: 1
        createdAt:
          type: string
          format: date-time

    WorkflowVersion:
      type: object
      description: Normalized workflow definition plus graph representation.
      required: [name, version, createdAt, dsl, graph]
      properties:
        name:
          type: string
        version:
          type: integer
          format: int32
          minimum: 1
        createdAt:
          type: string
          format: date-time
        dsl:
          $ref: '#/components/schemas/Dsl'
        graph:
          $ref: '#/components/schemas/Graph'

    Dsl:
      type: object
      description: Normalized workflow definition.
      required: [name, steps]
      properties:
        name:
          type: string
        steps:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/DslStep'

    DslStep:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
          description: Unique step identifier within the workflow.
          example: validate
        type:
          type: string
          description: Step type.
          enum: [noop, http]
        dependsOn:
          type: array
          description: List of step IDs that must complete before this step can run.
          items:
            type: string
          default: []

    Graph:
      type: object
      description: UI-friendly graph representation (nodes + edges).
      required: [steps, edges]
      properties:
        steps:
          type: array
          items:
            $ref: '#/components/schemas/GraphStep'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'

    GraphStep:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string

    GraphEdge:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
        to:
          type: string

    TenantResponse:
      type: object
      description: Tenant metadata.
      required: [tenantId, slug, name, createdAt]
      properties:
        tenantId:
          type: string
        slug:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          nullable: true

    ListTenantsResponse:
      type: object
      description: List of tenants.
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TenantResponse'

    CreateTenantRequest:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
          description: Stable tenant slug.
          minLength: 1
          maxLength: 64
          example: acme
        name:
          type: string
          description: Human-readable tenant name.
          minLength: 1
          maxLength: 128
          example: Acme Corp
        createCreatorMembership:
          type: boolean
          default: true
          description: |
            If true, the authenticated creator is added as the first tenant ADMIN.

    CreateTenantMembershipRequest:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
          description: Stable user identifier (for example, Supabase sub).
        role:
          type: string
          enum: [ADMIN, OPERATOR, VIEWER]

    TenantMembershipResponse:
      type: object
      required: [tenantId, userId, role, createdAt]
      properties:
        tenantId:
          type: string
        userId:
          type: string
        role:
          type: string
          enum: [ADMIN, OPERATOR, VIEWER]
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          nullable: true

    ApiError:
      type: object
      description: Standard error response.
      required: [code, message]
      properties:
        code:
          type: string
          example: WORKFLOW_INVALID_DAG
        messages:
          type: array
          items:
            type: string